+++ gsa/src/web/app.js
@@ -33,7 +33,7 @@ import ErrorBoundary from 'web/components/error/errorboundary';
 
 import LocaleObserver from 'web/components/observer/localeobserver';
 
-import GmpProvider from 'web/components/provider/gmpprovider';
+import GmpContext from 'web/components/provider/gmpprovider';
 
 import {
   setUsername,
@@ -99,13 +99,13 @@ class App extends React.Component {
   render() {
     return (
       <ErrorBoundary message={_('An error occurred on this page')}>
-        <GmpProvider gmp={gmp}>
+        <GmpContext.Provider value={gmp}>
           <StoreProvider store={store}>
             <LocaleObserver>
               <Routes />
             </LocaleObserver>
           </StoreProvider>
-        </GmpProvider>
+        </GmpContext.Provider>
       </ErrorBoundary>
     );
   }